// src/api/openapi/auto-generated/models/DisclosureType.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const DisclosureType = {
    chronic: 'chronic',
    acute: 'acute',
    unknownDefaultOpenApi: '11184809'
} as const;
export type DisclosureType = typeof DisclosureType[keyof typeof DisclosureType];


export function instanceOfDisclosureType(value: any): boolean {
    for (const key in DisclosureType) {
        if (Object.prototype.hasOwnProperty.call(DisclosureType, key)) {
            if ((DisclosureType as Record<string, DisclosureType>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function DisclosureTypeFromJSON(json: any): DisclosureType {
    return DisclosureTypeFromJSONTyped(json, false);
}

export function DisclosureTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): DisclosureType {
    return json as DisclosureType;
}

export function DisclosureTypeToJSON(value?: DisclosureType | null): any {
    return value as any;
}